GenAI/RAGs/project/streamlit_app/requirements.txt
streamlit
transformers
torch
faiss-cpu
sentencepiece
python-dotenv
